rollback: getCommentSize
