+++ src/comments/FlashComment.ts
@@ -12,6 +12,7 @@ import {
   getConfig,
   getFlashFontIndex,
   getFlashFontName,
+  getStrokeColor,
   isLineBreakResize,
   nativeSort,
   parseCommandAndNicoScript,
@@ -41,6 +42,42 @@ class FlashComment extends BaseComment {
     return this.getCommentSize(this.parseCommandAndNicoscript(comment));
   }
 
+  /**
+   * コメントの描画サイズを計算する
+   */
+  override getCommentSize(
+    parsedData: formattedCommentWithFont
+  ): formattedCommentWithSize {
+    this.context.font = parseFont(parsedData.font, parsedData.fontSize);
+    const size = parsedData as formattedCommentWithSize;
+    if (parsedData.invisible) {
+      size.height = 0;
+      size.width = 0;
+      size.lineHeight = 0;
+      size.fontSize = 0;
+      size.content = [];
+      size.resized = false;
+      size.resizedX = false;
+      size.resizedY = false;
+      size.charSize = 0;
+      return size;
+    }
+    const measure = this.measureText(parsedData);
+    if (options.scale !== 1 && size.layer === -1) {
+      measure.height *= options.scale;
+      measure.width *= options.scale;
+    }
+    size.height = measure.height * this._globalScale;
+    size.width = measure.width * this._globalScale;
+    size.lineHeight = measure.lineHeight;
+    size.fontSize = measure.fontSize;
+    size.content = measure.content;
+    size.resized = measure.resized;
+    size.resizedX = measure.resizedX;
+    size.resizedY = measure.resizedY;
+    size.charSize = measure.charSize;
+    return size;
+  }
   /**
    * コメントに含まれるニコスクリプトを処理する
    * @param comment
@@ -295,6 +332,10 @@ class FlashComment extends BaseComment {
     const { image, context } = this.createCanvas();
     image.width = this.comment.width;
     image.height = this.comment.height;
+    context.strokeStyle = getStrokeColor(this.comment);
+    context.fillStyle = this.comment.color;
+    context.textAlign = "start";
+    context.textBaseline = "alphabetic";
     context.lineWidth = 4;
     context.font = parseFont(this.comment.font, this.comment.fontSize);
     const scale =
